requirements.txt
Streamlit
Pillow